+++ lib/bsearch.h
@@ -48,11 +48,11 @@ extern int bsearch_mem_mbox(const char *word,
                             unsigned long hint,
                             unsigned long *linelenp);
 
-extern int bsearch_ncompare_mbox(const char *s1, int l1, const char *s2, int l2);
+extern int bsearch_ncompare_mbox(const char *s1, size_t l1, const char *s2, size_t l2);
 extern int bsearch_memtree_mbox(const unsigned char *s1, size_t l1,
                                 const unsigned char *s2, size_t l2);
 
-extern int bsearch_ncompare_raw(const char *s1, int l1, const char *s2, int l2);
+extern int bsearch_ncompare_raw(const char *s1, size_t l1, const char *s2, size_t l2);
 
 extern int cmpstringp_raw(const void *p1, const void *p2);
 extern int cmpstringp_mbox(const void *p1, const void *p2);
